+++ supvisors/tests/test_process.py
@@ -260,7 +260,7 @@ def test_process_create(supvisors_instance):
     assert process.forced_state is None
     assert process.forced_reason == ''
     assert process.expected_exit
-    assert process.last_event_time == 0
+    assert process.last_event_mtime == 0
     assert process.extra_args == ''
     assert process.running_identifiers == set()
     assert process.info_map == {}
@@ -676,16 +676,18 @@ def test_extra_args(mocker, supvisors_instance):
     assert process.extra_args == 'another args'
 
 
-def test_serialization(supvisors_instance):
+def test_serialization(mocker, supvisors_instance):
     """ Test the serialization of the ProcessStatus. """
+    mocker.patch('time.monotonic', return_value=1234.56)
     # test with a STOPPED process
     info = any_process_info_by_state(ProcessStates.STOPPED)
     process = create_process(info, supvisors_instance)
     process.add_info('10.0.0.1', info)
     serialized = process.serial()
     assert serialized == {'application_name': info['group'], 'process_name': info['name'],
+                          'now_monotonic': 1234.56, 'last_event_mtime': process.last_event_mtime,
                           'statecode': 0, 'statename': 'STOPPED', 'expected_exit': info['expected'],
-                          'last_event_time': process.last_event_time, 'identifiers': [], 'extra_args': ''}
+                          'identifiers': [], 'extra_args': ''}
     # test that returned structure is serializable using pickle
     dumped = pickle.dumps(serialized)
     loaded = pickle.loads(dumped)
@@ -697,8 +699,9 @@ def test_serialization(supvisors_instance):
     assert process._state == ProcessStates.STOPPED
     serialized = process.serial()
     assert serialized == {'application_name': info['group'], 'process_name': info['name'],
+                          'now_monotonic': 1234.56, 'last_event_mtime': process.last_event_mtime,
                           'statecode': 200, 'statename': 'FATAL', 'expected_exit': info['expected'],
-                          'last_event_time': process.last_event_time, 'identifiers': [], 'extra_args': ''}
+                          'identifiers': [], 'extra_args': ''}
 
 
 def test_get_applicable_details(supvisors_instance):
@@ -761,9 +764,9 @@ def test_add_info(supvisors_instance):
     process.extra_args = 'something else'
     process.add_info('10.0.0.1', info)
     # check last event info
-    assert process.last_event_time > 0.0
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime > 0.0
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check contents
     assert len(process.info_map) == 1
@@ -792,9 +795,9 @@ def test_add_info(supvisors_instance):
     info['expected'] = False
     process.add_info('10.0.0.1', info)
     # check last event info
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check contents
     assert len(process.info_map) == 1
@@ -813,9 +816,9 @@ def test_add_info(supvisors_instance):
     info = any_process_info_by_state(ProcessStates.RUNNING)
     process.add_info('10.0.0.2', info)
     # check last event info
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check contents
     assert len(process.info_map) == 2
@@ -834,9 +837,9 @@ def test_update_info(supvisors_instance):
     process = create_process(info, supvisors_instance)
     process.add_info('10.0.0.1', info)
     # test last event info stored
-    assert process.last_event_time > 0
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime > 0
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check changes on status
     assert process.info_map['10.0.0.1']['state'] == ProcessStates.STOPPED
@@ -849,9 +852,9 @@ def test_update_info(supvisors_instance):
     process.update_info('10.0.0.1', {'state': ProcessStates.STARTING, 'extra_args': '-x dummy',
                                      'now': 10, 'now_monotonic': 5})
     # test last event info stored
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check changes on status
     info = process.info_map['10.0.0.1']
@@ -870,9 +873,9 @@ def test_update_info(supvisors_instance):
     process.update_info('10.0.0.1', {'state': ProcessStates.RUNNING, 'now': 15, 'now_monotonic': 10,
                                      'pid': 1234, 'extra_args': '-z another'})
     # test last event info stored
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check changes
     assert info['state'] == ProcessStates.RUNNING
@@ -901,9 +904,9 @@ def test_update_info(supvisors_instance):
     info = any_process_info_by_state(ProcessStates.STOPPED)
     process.add_info('10.0.0.2', info)
     # test last event info stored
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     assert process.state == ProcessStates.RUNNING
     assert process.displayed_state == ProcessStates.FATAL
@@ -917,9 +920,9 @@ def test_update_info(supvisors_instance):
     process.update_info('10.0.0.2', {'state': ProcessStates.RUNNING, 'now': 25, 'now_monotonic': 20,
                                      'pid': 4321, 'extra_args': ''})
     # test last event info stored
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check state and addresses
     assert process.state == ProcessStates.RUNNING
@@ -932,9 +935,9 @@ def test_update_info(supvisors_instance):
     process.update_info('10.0.0.1', {'state': ProcessStates.EXITED, 'now': 30, 'now_monotonic': 25,
                                      'expected': False, 'extra_args': ''})
     # test last event info stored
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check changes
     assert info['state'] == ProcessStates.EXITED
@@ -955,9 +958,9 @@ def test_update_info(supvisors_instance):
     process.update_info('10.0.0.2', {'state': ProcessStates.STOPPING, 'now': 25, 'now_monotonic': 30,
                                      'extra_args': ''})
     # test last event info stored
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check changes
     assert info['state'] == ProcessStates.STOPPING
@@ -977,9 +980,9 @@ def test_update_info(supvisors_instance):
     process.update_info('10.0.0.2', {'state': ProcessStates.STOPPED, 'now': 26, 'now_monotonic': 35,
                                      'extra_args': ''})
     # test last event info stored
-    assert process.last_event_time >= last_event_time
-    last_event_time = process.last_event_time
-    assert last_event_time == info['local_time']
+    assert process.last_event_mtime >= last_event_time
+    last_event_time = process.last_event_mtime
+    assert last_event_time == info['local_mtime']
     assert info['event_time'] == info['now_monotonic']
     # check changes
     assert info['state'] == ProcessStates.STOPPED
@@ -1131,14 +1134,14 @@ def test_update_status(supvisors_instance):
     info = any_process_info_by_state(ProcessStates.FATAL)
     process = create_process(info, supvisors_instance)
     process.add_info('10.0.0.3', info)
-    process.info_map['10.0.0.3']['local_time'] = 10
+    process.info_map['10.0.0.3']['local_mtime'] = 10
     assert process.running_identifiers == set()
     assert process.state == ProcessStates.FATAL
     assert not process.expected_exit
     # add a STOPPED process info
     info = any_process_info_by_state(ProcessStates.STOPPED)
     process.add_info('10.0.0.1', info)
-    process.info_map['10.0.0.1']['local_time'] = 5
+    process.info_map['10.0.0.1']['local_mtime'] = 5
     process.update_status('10.0.0.1', ProcessStates.STOPPED)
     assert process.running_identifiers == set()
     assert process.state == ProcessStates.FATAL  # FATAL info above is more recent
@@ -1146,7 +1149,7 @@ def test_update_status(supvisors_instance):
     # replace with an EXITED process info
     info = any_process_info_by_state(ProcessStates.EXITED)
     process.update_info('10.0.0.3', info)
-    process.info_map['10.0.0.3']['local_time'] = 15
+    process.info_map['10.0.0.3']['local_mtime'] = 15
     process.update_status('10.0.0.3', ProcessStates.EXITED)
     assert process.running_identifiers == set()
     assert process.state == ProcessStates.EXITED
@@ -1154,7 +1157,7 @@ def test_update_status(supvisors_instance):
     # add a STARTING process info
     info = any_process_info_by_state(ProcessStates.STARTING)
     process.add_info('10.0.0.2', info)
-    process.info_map['10.0.0.2']['local_time'] = 20
+    process.info_map['10.0.0.2']['local_mtime'] = 20
     process.update_status('10.0.0.2', ProcessStates.STARTING)
     assert process.running_identifiers == {'10.0.0.2'}
     assert process.state == ProcessStates.STARTING
@@ -1162,7 +1165,7 @@ def test_update_status(supvisors_instance):
     # replace a BACKOFF process info
     info = any_process_info_by_state(ProcessStates.BACKOFF)
     process.update_info('10.0.0.3', info)
-    process.info_map['10.0.0.3']['local_time'] = 20
+    process.info_map['10.0.0.3']['local_mtime'] = 20
     process.update_status('10.0.0.3', ProcessStates.BACKOFF)
     assert process.running_identifiers == {'10.0.0.3', '10.0.0.2'}
     assert process.state == ProcessStates.BACKOFF
@@ -1170,7 +1173,7 @@ def test_update_status(supvisors_instance):
     # replace STARTING process info with RUNNING
     info = any_process_info_by_state(ProcessStates.RUNNING)
     process.update_info('10.0.0.2', info)
-    process.info_map['10.0.0.2']['local_time'] = 25
+    process.info_map['10.0.0.2']['local_mtime'] = 25
     process.update_status('10.0.0.2', ProcessStates.RUNNING)
     assert process.running_identifiers == {'10.0.0.3', '10.0.0.2'}
     assert process.state == ProcessStates.RUNNING
@@ -1178,7 +1181,7 @@ def test_update_status(supvisors_instance):
     # replace BACKOFF process info with FATAL
     info = any_process_info_by_state(ProcessStates.FATAL)
     process.update_info('10.0.0.3', info)
-    process.info_map['10.0.0.3']['local_time'] = 25
+    process.info_map['10.0.0.3']['local_mtime'] = 25
     process.update_status('10.0.0.3', ProcessStates.FATAL)
     assert process.running_identifiers == {'10.0.0.2'}
     assert process.state == ProcessStates.RUNNING
@@ -1187,7 +1190,7 @@ def test_update_status(supvisors_instance):
     # in ProcessInfoDatabase, EXITED processes have a stop date later than STOPPED processes
     info = any_process_info_by_state(ProcessStates.STOPPED)
     process.update_info('10.0.0.2', info)
-    process.info_map['10.0.0.2']['local_time'] = 30
+    process.info_map['10.0.0.2']['local_mtime'] = 30
     process.update_status('10.0.0.2', ProcessStates.STOPPED)
     assert not process.running_identifiers
     assert process.state == ProcessStates.STOPPED